feat: compliated 9 video on course
