+++ first-project/.eslintcache
@@ -1,1 +1,1 @@
-[{"E:\\study\\learn_react\\first-project\\src\\reportWebVitals.js":"1","E:\\study\\learn_react\\first-project\\src\\App.js":"2"},{"size":362,"mtime":499162500000,"results":"3","hashOfConfig":"4"},{"size":498,"mtime":1606258333684,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"mbdj42",{"filePath":"8","messages":"9","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"E:\\study\\learn_react\\first-project\\src\\reportWebVitals.js",[],"E:\\study\\learn_react\\first-project\\src\\App.js",["10","11","12","13"],{"ruleId":"14","severity":1,"message":"15","line":1,"column":8,"nodeType":"16","messageId":"17","endLine":1,"endColumn":12},{"ruleId":"18","severity":1,"message":"19","line":16,"column":7,"nodeType":"20","endLine":16,"endColumn":19},{"ruleId":"18","severity":1,"message":"19","line":17,"column":7,"nodeType":"20","endLine":17,"endColumn":19},{"ruleId":"18","severity":1,"message":"19","line":18,"column":7,"nodeType":"20","endLine":18,"endColumn":19},"no-unused-vars","'logo' is defined but never used.","Identifier","unusedVar","jsx-a11y/anchor-is-valid","The href attribute requires a valid value to be accessible. Provide a valid, navigable address as the href value. If you cannot provide a valid href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement"]
+[{"E:\\study\\learn_react\\first-project\\src\\reportWebVitals.js":"1","E:\\study\\learn_react\\first-project\\src\\App.js":"2","E:\\study\\learn_react\\first-project\\src\\index.js":"3","E:\\study\\learn_react\\first-project\\src\\Header.js":"4","E:\\study\\learn_react\\first-project\\src\\Technologies.js":"5"},{"size":362,"mtime":499162500000,"results":"6","hashOfConfig":"7"},{"size":243,"mtime":1606296147676,"results":"8","hashOfConfig":"7"},{"size":500,"mtime":499162500000,"results":"9","hashOfConfig":"7"},{"size":211,"mtime":1606290367793,"results":"10","hashOfConfig":"7"},{"size":248,"mtime":1606290360924,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"mbdj42",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\study\\learn_react\\first-project\\src\\reportWebVitals.js",[],"E:\\study\\learn_react\\first-project\\src\\App.js",[],"E:\\study\\learn_react\\first-project\\src\\index.js",[],"E:\\study\\learn_react\\first-project\\src\\Header.js",["22","23","24"],"E:\\study\\learn_react\\first-project\\src\\Technologies.js",[],{"ruleId":"25","severity":1,"message":"26","line":6,"column":7,"nodeType":"27","endLine":6,"endColumn":19},{"ruleId":"25","severity":1,"message":"26","line":7,"column":7,"nodeType":"27","endLine":7,"endColumn":19},{"ruleId":"25","severity":1,"message":"26","line":8,"column":7,"nodeType":"27","endLine":8,"endColumn":19},"jsx-a11y/anchor-is-valid","The href attribute requires a valid value to be accessible. Provide a valid, navigable address as the href value. If you cannot provide a valid href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement"]
