+++ first-project/src/Header.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <div>
+      <a href="#">Home</a>
+      <a href="#">About</a>
+      <a href="#">Contacts</a>
+    </div>
+  );
+}
+
+export default Header;
