+++ first-project/src/Technologies.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Technologies = () => {
+  return (
+    <div>
+      <ul>
+        <li>html</li>
+        <li>css</li>
+        <li>js</li>
+        <li>react</li>
+      </ul>
+    </div>
+  )
+}
+
+export default Technologies;
